+++ CHANGELOG.md
@@ -1,12 +1,25 @@
 ## [Version - 0]
+
+## Fix [v0.39.4] 2022-11-1
+
+- fixed text color in new task menu
+- fixed status mission menu
+- close pin in add/edit news menu
+
 ## Fix [v0.39.3] 2022-10-31
+
 - fixed field time news
+
 ## Fix [v0.39.2] 2022-10-31
+
 - fixed style
+
 ## Fix [v0.39.1] 2022-10-30
+
 - import css for news guru
 
 ## Feature [v0.39.0] 2022-10-30
+
 - eslint & prettier setup
 - lazy load
 - Pin News Screen